requirements.txt
keras==3.2.1
numpy==1.26.4
Pillow==9.4.0
Pillow==9.5.0
Pillow==10.3.0
plotly==5.9.0
Requests==2.31.0
streamlit==1.33.0
tensorflow==2.16.1
tensorflow_intel==2.16.1
